requirements.txt
telethon
asyncio
aiohttp
speedtest-cli
